# backend-api/openapi/openapi.yaml
components:
...
  schemas:
...
    PersonRef:
      properties:
        lastName:
          type: string
        name:
          type: string
        personId:
          format: int32
          type: integer
      type: object
...
    SubstitutionRequest:
      properties:
        dateEnd:
          format: date-time
          type: string
        dateStart:
          format: date-time
          type: string
        fkIdSubstitute:
          $ref: '#/components/schemas/PersonRef'
        fkIdSubstituted:
          $ref: '#/components/schemas/PersonRef'
        notes:
          type: string
        substitutePersonId:
          format: int32
          type: integer
        substitutedPersonId:
          format: int32
          type: integer
      required:
      - dateEnd
      - dateStart
      type: object
    SubstitutionResponse:
      properties:
        dateEnd:
          format: date-time
          type: string
        dateStart:
          format: date-time
          type: string
        fkIdSubstitute:
          $ref: '#/components/schemas/PersonRef'
        fkIdSubstituted:
          $ref: '#/components/schemas/PersonRef'
        idSubstitution:
          format: int32
          type: integer
        notes:
          type: string
        substitutePersonId:
          format: int32
          type: integer
        substitutedPersonId:
          format: int32
          type: integer
        version:
          format: int32
          type: integer
      type: object
...
info:
  contact:
    email: info@licensis.com
    name: Licensis
    url: https://www.licensis.com
  description: Sistema de gestión notarial — API REST backend. Todos los recursos siguen el prefijo `/api/v1/`. Las respuestas
    de error siguen el esquema `{error, message, timestamp, path}`.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Notaire API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/suplencia:
    get:
      operationId: getAll_7
      responses:
        '200':
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/SubstitutionResponse'
                type: array
          description: OK
      summary: Obtener todos los suplencia
      tags:
      - Suplencia
    post:
      operationId: create_9
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubstitutionRequest'
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                type: object
          description: Creado
        '400':
          content:
            '*/*':
              schema:
                type: object
          description: Solicitud inválida
        '409':
          content:
            '*/*':
              schema:
                type: object
          description: Conflicto
      summary: Crear nuevo suplencia
      tags:
      - Suplencia
  /api/v1/suplencia/{id}:
    delete:
      operationId: delete_9
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '204':
          description: Eliminado
        '404':
          description: No encontrado
      summary: Eliminar suplencia
      tags:
      - Suplencia
    get:
      operationId: getById_9
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SubstitutionResponse'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SubstitutionResponse'
          description: No encontrado
      summary: Obtener suplencia por ID
      tags:
      - Suplencia
    put:
      operationId: update_9
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubstitutionRequest'
        required: true
      responses:
        '200':
          description: OK
        '404':
          description: No encontrado
      summary: Actualizar suplencia
      tags:
      - Suplencia
...
servers:
- description: Desarrollo local
  url: http://localhost:8080
- description: Docker local
  url: http://localhost:8081
...